Display a list of top extensions and skins
Now that we have some statistics of what people are downloading, lets
expose them in the user interface!

For now the list is hardcoded in configuration, at some point we should
automate it. But this exposes it in the frontend for now.

Since the styles are now needed in no-JS mode, move them to a separate
module and load it with addModuleStyles().

Bug: T120339
Change-Id: Ifafa5a317b1b34d693f1ef4925d6e8449a3c560a