Retrieve PopularList directly from graphite if configured
This change introduces a new setting $wgExtDistGraphiteRenderApi
which should point to a render URL for a graphite
instance that the installation of MediaWiki sends statsd
data to. On Wikimedia sites, this will be set to
"https://graphite.wikimedia.org/render".

On special page load a graphite call will be made
requesting the top downloads of the given type.

This list will be cached for 1 day.

If no graphite API is configured we fall back to the
popular list globals. And if those are not set / empty
we don't display a list.

Bug: T120340
Bug: T120339
Change-Id: I00e59daaf6060371244b6b7d429f9d639138fbd7